+++ src/test/kotlin/lotto/domain/LottoGeneratorTest.kt
@@ -1,8 +1,7 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import lotto.domain.LottoGenerator
 
 class LottoGeneratorTest : BehaviorSpec({
     Given("로또를 생성하면") {
