+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -1,11 +1,10 @@
-package lotto
+package lotto.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.data.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
-import lotto.domain.Lotto
 
 class LottoTest : BehaviorSpec({
     Given("유효한 로또 번호로 생성") {
